🚚 #94 move all phaser logic in a separate folder
