+++ src/components/GamePhaser/index.tsx
@@ -2,11 +2,9 @@ import React, { FC, useRef, useEffect, useCallback, useState } from "react";
 import * as Phaser from "phaser";
 import { selectors } from "../../redux";
 import { useDispatch, useSelector } from "react-redux";
-import PhaserMatterCollisionPlugin from "phaser-matter-collision-plugin";
-
+import { getGameConfig } from "../../phaser/configs/gameConfig";
+import MainScene from "../../phaser/scenes/MainScene";
 import "./index.scss";
-
-import MainScene from "./MainScene";
 
 interface Props {
   isActive: boolean;
@@ -53,50 +51,12 @@ const GamePhaser: FC<Props> = ({ isActive }) => {
   // create phaser game
 
   const createGame = useCallback(() => {
-    const gameConfig: Phaser.Types.Core.GameConfig = {
-      type: Phaser.AUTO,
-      width: areaWidth,
-      height: areaHeight,
-      parent: $parent.current!,
-      transparent: true,
-      physics: {
-        default: "matter",
-        matter: {
-          gravity: { y: 1 },
-          debug:
-            process.env.NODE_ENV === "development"
-              ? {
-                  showBody: true,
-                  showStaticBody: true,
-                  showInternalEdges: true,
-                  showVelocity: true,
-                  showCollisions: true,
-                  lineThickness: 2,
-                  showPositions: true,
-                  positionSize: 6
-                }
-              : false,
-          setBounds: {
-            x: -150,
-            left: false,
-            right: false,
-            width: areaWidth + 150,
-            thickness: 30
-          },
-          "plugins.wrap": true
-        }
-      },
-      plugins: {
-        scene: [
-          {
-            plugin: PhaserMatterCollisionPlugin, // The plugin class
-            key: "matterCollision",
-            mapping: "matterCollision"
-          }
-        ]
-      },
-      scene: $mainScene.current
-    };
+    const gameConfig = getGameConfig({
+      areaHeight,
+      areaWidth,
+      scene: $mainScene.current,
+      parent: $parent.current!
+    });
 
     $game.current = new Phaser.Game(gameConfig);
